Shockwaves

Documentation

class Display a where

Class determining the appearance of a value in the waveform viewer (text and color). By default, this uses Show and VKNormal.

Minimal complete definition

Nothing

Methods

display :: a -> (ValueRepr, ValueKind)

repr :: a -> ValueRepr

default repr :: Show a => a -> ValueRepr

kind :: a -> ValueKind

data ValueRepr

Representation of a value.

Constructors

VRBit Char
Single bit

VRBits String
Bitvector

VRString String
Text

VRNotPresent
Signal not present; do not display signal at all

data ValueKind

Determines the way values are displayed. For most signals, this only determines the color, but VIBool signals can have lines at different heights for the different value types.

Constructors

VKNormal
Green

VKUndef
Red

VKHighImp
Yellow

VKCustom Color
Any

VKWarn
Red

VKDontCare
Blue

VKWeak
Gray

class Split a where

Class for determining the structure and value of subsignals. The structure can be automatically deduced for types implementing Generic with all subtypes implementing Split as well.

Minimal complete definition

Nothing

Methods

structure :: VariableInfo

The structure for the signal. Only VICompound types allow for subsignals. In addition to the the structure of subsignals, this also controls the way the current signal is displayed. Most data will be shown as blocks, but boolean types are displayed as a single line that can be high or low (or in between for special values).

default structure :: forall x. (Generic a, AutoSplit (Rep a x)) => VariableInfo

split :: a -> (ValueRepr, ValueKind) -> [STR]

Function to create the data for subsignals of a given type. If structure is not of the VICompound type, this list must be empty. Subsignals need to share the names and order used in structure, but not all subsignals need to be provided. Subsignals that are left out will be automatically set to VRNotPresent. The function is given a copy of the display value, in case this needs to be copied.

default split :: forall x. (Generic a, AutoSplit (Rep a x)) => a -> (ValueRepr, ValueKind) -> [STR]

data VariableInfo

Information about the signal structure. VICompound is the only variant allowed to have subsignals. Translation results must match this structure. VIBool and VIClock are displayed differently in the waveform viewer (as waves).

Constructors

VICompound [(String, VariableInfo)]
A signal with subsignals

VIBits
A multiple bit signal

VIBool
A 1-bit signal; displayed as a proper wave

VIClock
A clock signal; displayed as a proper wave

VIString
A simple string

VIReal
A real number

translate :: (Display a, Split a) => a -> TranslationResult

Translate a value using safeDisplay and safeSplit. If either returns Nothing, return a translation representing undefined instead.