Shockwaves.PostTranslation

Description

Module for translating values after simulation (i.e. when not using Trace). This module is mostly useful for experimentation.

Translation tables

genTable :: (String -> (StructF, TransF)) -> [(String, [String])] -> TranslationTable

Generate a table of value representations for types, using the provided type-label-to-functions table, and the list of (type label, list of values) pairs.

type StructF

Arguments

= VariableInfo
StructF is treated as a function returning the variable info.

type TransF

Arguments

= String -> TranslationResult
Function translating a string of bits into a TranslationResult.

type TranslationTable = Map String (VariableInfo, Map String TranslationResult)

Table holding for each type label the structure of the type, and the translations of some values.

class TypeFunctions a where

Get translation function for a given type. This class is used to only have to specify the type once while translating.

Methods

tf :: (StructF, TransF)

Get translation function for the type.

Files and command line

translateFile :: (String -> (StructF, TransF)) -> String -> String -> IO ()

Given a function to convert type strings into translation functions for that type, turn the values per type from the input file into a translation table.

translateCmdLine :: (String -> (StructF, TransF)) -> IO ()

Given a function that maps string representations of types to the translation functions for said type, translate all values in the input file. Run translateFile using the first two command line arguments for the input and output. The input file format is:

<type>
<value> <value> <value> ...
<type>
...